+++ src/main/java/no/taardal/pixelcave/state/actor/knight/KnightAttackingMidAirState.java
@@ -12,7 +12,7 @@ public class KnightAttackingMidAirState extends KnightAttackingState {
     }
 
     @Override
-    public void update(double secondsSinceLastUpdate) {
+    public void update(float secondsSinceLastUpdate) {
         if (getAnimation().isFinished()) {
             stateMachine.popState();
         } else if (!enemiesAttacked && getAnimation().getFrame() == 1) {
